requirements.txt
# Environment & Utils
python-dotenv
tqdm
click
tyro
pyyaml
rich

# Core Processing (ASR, LLM)
faster-whisper
google-generativeai
openai


# PyTorch Core (Platform dependent installation recommended, but listing basics)
torch>=2.0.0
torchaudio
torchvision

# Data & Math
numpy<2.0
scipy
matplotlib

# Audio Processing
librosa>=0.9.0,<0.11.0
numba>=0.55.0  # librosa dependency
soundfile
pydub
pypinyin
jieba
vocos

# Visual Processing
opencv-python
pillow
imageio
imageio-ffmpeg
moviepy
scikit-image
albumentations

# ML/DL Frameworks & Extras
transformers
accelerate
gradio>=5.0.0
hydra-core
ema_pytorch
cached_path
wandb
safetensors
onnx
onnxruntime

# F5-TTS 特定依賴
datasets
x-transformers>=1.31.14
rjieba
unidecode
torchdiffeq
tomli
# transformers_stream_generator  # 版本不兼容，僅訓練需要
# torchcodec  # Mac M4 無法使用，F5-TTS 不需要（已卸載）✅

# Web UI
streamlit==1.30.0

# Face detection（如果尚未安裝）
face-alignment

# 時間軸處理
pyrubberband  # 需要先 brew install rubberband
audiotsm      # 備選方案

# OCR - 使用 EasyOCR (Mac M4 友好)
easyocr
# 或使用 PaddleOCR:
# paddlepaddle>=3.0.0
# paddleocr

# Inpainting (如果選擇 IOPaint)
iopaint  # 或不裝，用 OpenCV 的 cv2.inpaint